requirements.txt
fastapi
uvicorn
openai
edge-tts
sqlalchemy
pydantic
python-dotenv
python-multipart
httpx
